third_party/blink/web_tests/external/wpt/wai-aria/role/invalid-roles-expected.txt

This is a testharness.js-based test.
Found 76 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] nav with invalid role name foo
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with two invalid role names
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with three invalid role names
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] button with invalid role name foo
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] button with two invalid role names
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] button with three invalid role names
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with empty character as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with line break (\r) character as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with tab (	) as role (should be treated as whitespace)
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with zero-width space as role (should be treated as whitespace)
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with braille space (10240) as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with braille space (x2800) as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with non-breaking space (nbsp) as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with standard space (nbsp) as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] link with role set to tilde diacritic
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] link with role foo that has tilde diacritic
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] link with role set to overline diacritic
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] link with role foo that has overline diacritic
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] link with role set to circumflex diacritic below
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] link with role foo that has circumflex diacritic below
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] link with role set to button with tilde diacritic
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] button with role set to zero-width joiner
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] button with role set to foo with zero-width joiner
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with role set to theta (Θ)
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with role set to Persian character (ژ)
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with multiple non-latin character roles, Persian character (ژ) and ♮
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with role set to Japanese kanji
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with role set to Arabic text
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with multiple role assignments set to Arabic text
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with role set to Hebrew text
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with multiple role assignments set to Hebrew text
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] link with role set to ampersand character
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] link with role set to less than angle bracket character
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with role set to region followed by backslash
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with role set to backslash followed by region
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] nav with role set to region with backslash after e character
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with invalid role name foo
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with two invalid role names
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with three invalid role names
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] div with invalid role name foo
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] div with two invalid role names
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] div with three invalid role names
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with escaped empty character as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with escaped line break (\r) character as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with escaped tab (	) as role (should be treated as whitespace)
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with escaped zero-width space as role (should be treated as whitespace)
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with escaped braille space (10240) as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with escaped braille space (x2800) as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with escaped non-breaking space (nbsp) as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with escaped standard space (nbsp) as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with empty character as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with line break (\r) character as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with tab (	) as role (should be treated as whitespace)
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with zero-width space as role (should be treated as whitespace)
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with braille space (10240) as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with non-breaking space (nbsp) as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with standard space as role
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] div with role set to tilde diacritic
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] div with role foo that has tilde diacritic
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] div with role set to button with tilde diacritic
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] div with role set to button with unescaped tilde diacritic
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with role set to theta (Θ)
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with role set to Persian character (ژ)
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with multiple non-latin character roles, Persian character (ژ) and ♮
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with role set to Japanese kanji
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] div with role set to Arabic text
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] div with multiple role assignments set to Arabic text
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] div with role set to Hebrew text
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] div with multiple role assignments set to Hebrew text
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with role set to ampersand character
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with role set to less than angle bracket character
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with role set to unescaped ampersand character
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with role set to unescaped less than angle bracket character
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with role set to region followed by backslash
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with role set to backslash followed by region
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
[FAIL] span with role set to region with forward slash after e character
  promise_test: Unhandled rejection with value: object "TypeError: window.test_driver_internal.get_computed_role is not a function"
Harness: the test ran to completion.
